+++ CommonConcepts/Plugins/Rhetos.CommonConcepts.Test/GenericFilterHelperTest.cs
@@ -74,7 +74,7 @@ namespace Rhetos.CommonConcepts.Test
             var genericFilter = new FilterCriteria("Name", operation, value);
             Console.WriteLine("genericFilter: " + genericFilter.Property + " " + genericFilter.Operation + " " + genericFilter.Value);
 
-            var genericFilterHelper = new GenericFilterHelper(new DomainObjectModelMock(), new DataStructureReadParametersStub());
+            var genericFilterHelper = new GenericFilterHelper(new DomainObjectModelMock(), new DataStructureReadParametersStub(), new CommonConceptsRuntimeOptions(), new ConsoleLogProvider());
             var filterObject = genericFilterHelper.ToFilterObjects(typeof(C).FullName, new FilterCriteria[] { genericFilter }).Single();
             Console.WriteLine("filterObject.FilterType: " + filterObject.FilterType.FullName);
             var filterExpression = genericFilterHelper.ToExpression<C>((IEnumerable<PropertyFilter>)filterObject.Parameter);
@@ -200,10 +200,10 @@ namespace Rhetos.CommonConcepts.Test
             Assert.AreEqual("3", TestUtility.DumpSorted(TestFilter("RefID", "NotIn", new string[] { "", id1.ToString(), id2.ToString(), id2.ToString(), id4.ToString() }, items), item => item.Name));
         }
 
-        private IEnumerable<T> TestFilter<T>(string property, string operation, object value, IEnumerable<T> items)
+        private static IEnumerable<T> TestFilter<T>(string property, string operation, object value, IEnumerable<T> items)
         {
             var genericFilter = new FilterCriteria(property, operation, value);
-            var genericFilterHelper = new GenericFilterHelper(new DomainObjectModelMock(), new DataStructureReadParametersStub());
+            var genericFilterHelper = new GenericFilterHelper(new DomainObjectModelMock(), new DataStructureReadParametersStub(), new CommonConceptsRuntimeOptions(), new ConsoleLogProvider());
             var filterObject = genericFilterHelper.ToFilterObjects(typeof(T).FullName, new FilterCriteria[] { genericFilter }).Single();
             var filterExpression = genericFilterHelper.ToExpression<T>((IEnumerable<PropertyFilter>)filterObject.Parameter);
             var filteredItems = items.AsQueryable().Where(filterExpression).ToList();